init: start of state management task, refactor file structure
