+++ package.json
@@ -41,6 +41,7 @@
     "husky": "^9.1.7",
     "jsdom": "^26.0.0",
     "prettier": "3.4.2",
+    "sass-embedded": "^1.83.4",
     "typescript": "~5.6.2",
     "typescript-eslint": "^8.18.2",
     "vite": "^6.0.5",
