+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { Route, Routes, Navigate } from 'react-router';
-import HomePage from './views/HomePage';
-import NotFound404 from './components/NotFound404';
-import PokemonCard from './components/PokemonCard';
+import HomePage from './views/Homepage/Homepage';
+import NotFound404 from './components/NotFound404/NotFound404';
+import PokemonCard from './components/PokemonCard/PokemonCard';
 
 export default function App() {
   return (
