+++ src/components/ErrorButton/ErrorButton.scss
@@ -1,0 +1,18 @@
+.error-button {
+  position: absolute;
+  right: 30px;
+  bottom: 30px;
+  height: 40px;
+  width: 150px;
+  font-size: 16px;
+  background-color: #dc3545;
+  color: white;
+  border: none;
+  border-radius: 5px;
+  cursor: pointer;
+  transition: background-color 0.3s;
+
+  &:hover {
+    background-color: #a71d2a;
+  }
+}
