+++ src/components/Loader/Loader.tsx
@@ -1,3 +1,5 @@
+import './Loader.scss';
+
 export default function Loader() {
   return (
     <div className="lds-roller">
