+++ src/components/NotFound404/NotFound404.scss
@@ -1,0 +1,4 @@
+.not-found {
+  display: flex;
+  flex-direction: column;
+}
