+++ src/components/NotFound404/NotFound404.tsx
@@ -1,8 +1,10 @@
+import './NotFound404.scss';
+
 export default function NotFound404() {
   return (
-    <>
+    <div className="not-found">
       <h1>404</h1>
       <h1>PAGE NOT FOUND</h1>
-    </>
+    </div>
   );
 }
