+++ src/components/Pagination/Pagination.scss
@@ -1,0 +1,4 @@
+.pagination {
+  display: flex;
+  gap: 20px;
+}
