+++ src/components/PokemonCard/PokemonCard.scss
@@ -1,0 +1,38 @@
+.pokemon-card {
+  padding: 30px;
+  border-radius: 20px;
+  border: 2px solid black;
+}
+
+.pokemon-card__close-button {
+  align-self: end;
+  width: 20px;
+  height: 20px;
+  padding: 5px;
+  text-decoration: none;
+  color: black;
+  border: 1px solid black;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  border-radius: 5px;
+}
+
+.pokemon-card__close-button:hover {
+  background-color: #ccc;
+}
+
+.pokemon-card__name {
+  text-transform: capitalize;
+}
+
+.pokemon-card__details {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.pokemon-card__images {
+  display: flex;
+  gap: 10px;
+}
