+++ src/components/PokemonList/PokemonList.scss
@@ -1,0 +1,34 @@
+.pokemon-list {
+  width: 60%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  gap: 20px;
+  text-transform: capitalize;
+}
+
+.pokemon-list__items {
+  display: grid;
+  grid-template-columns: repeat(5, 1fr);
+  grid-template-rows: repeat(4, 1fr);
+  gap: 10px;
+}
+
+.pokemon-list__name {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 150px;
+  height: 50px;
+  border: 2px solid black;
+  border-radius: 25px;
+  color: black;
+  text-decoration: none;
+  font-size: 20px;
+  font-weight: bold;
+}
+
+.pokemon-list__name:hover {
+  background-color: #ccc;
+}
