+++ src/components/SearchBox/SearchBox.scss
@@ -1,0 +1,39 @@
+.search-container {
+  margin-top: 20px;
+  display: flex;
+  justify-content: center;
+  gap: 10px;
+
+  input {
+    width: 20vw;
+    padding: 10px;
+    font-size: 16px;
+    border: 2px solid #ccc;
+    border-radius: 5px;
+  }
+}
+
+.search-button {
+  padding: 10px 15px;
+  font-size: 18px;
+  background-color: #007bff;
+  color: white;
+  border: none;
+  border-radius: 5px;
+  cursor: pointer;
+  text-decoration: none;
+  transition: background-color 0.3s;
+
+  &:hover {
+    background-color: #0056b3;
+  }
+}
+
+.current {
+  background-color: #0056b3;
+}
+
+.disabled {
+  pointer-events: none;
+  opacity: 0.5;
+}
