+++ src/index.css
@@ -10,247 +10,3 @@ main {
   display: flex;
   flex-direction: column;
 }
-
-/* Container for search box */
-.search-container {
-  margin-top: 20px;
-  display: flex;
-  justify-content: center;
-  gap: 10px;
-}
-
-.search-container input {
-  width: 20vw;
-  padding: 10px;
-  font-size: 16px;
-  border: 2px solid #ccc;
-  border-radius: 5px;
-}
-
-.search-button {
-  padding: 10px 15px;
-  font-size: 18px;
-  background-color: #007bff;
-  color: white;
-  border: none;
-  border-radius: 5px;
-  cursor: pointer;
-  text-decoration: none;
-  transition: background-color 0.3s;
-}
-
-.search-button:hover {
-  background-color: #0056b3;
-}
-
-.current {
-  background-color: #0056b3;
-}
-
-.disabled {
-  pointer-events: none;
-  opacity: 0.5;
-}
-
-/* Results styling */
-
-.results {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.pokemon-card {
-  padding: 30px;
-  border-radius: 20px;
-  border: 2px solid black;
-}
-
-.pokemon-card__close-button {
-  align-self: end;
-  width: 20px;
-  height: 20px;
-  padding: 5px;
-  text-decoration: none;
-  color: black;
-  border: 1px solid black;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  border-radius: 5px;
-}
-
-.pokemon-card__close-button:hover {
-  background-color: #ccc;
-}
-
-.pokemon-card__name {
-  text-transform: capitalize;
-}
-
-.pokemon-card__details {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.pokemon-card__images {
-  display: flex;
-  gap: 10px;
-}
-
-.pokemon-list {
-  width: 60%;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 20px;
-  text-transform: capitalize;
-}
-
-.pokemon-list__items {
-  display: grid;
-  grid-template-columns: repeat(5, 1fr);
-  grid-template-rows: repeat(4, 1fr);
-  gap: 10px;
-}
-
-.pokemon-list__name {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 150px;
-  height: 50px;
-  border: 2px solid black;
-  border-radius: 25px;
-  color: black;
-  text-decoration: none;
-  font-size: 20px;
-  font-weight: bold;
-}
-
-.pokemon-list__name:hover {
-  background-color: #ccc;
-}
-
-.pokemon-list__pagination {
-  display: flex;
-  gap: 20px;
-}
-
-/* Error button styling */
-.error-button {
-  position: absolute;
-  right: 30px;
-  bottom: 30px;
-  height: 40px;
-  width: 150px;
-  font-size: 16px;
-  background-color: #dc3545;
-  color: white;
-  border: none;
-  border-radius: 5px;
-  cursor: pointer;
-  transition: background-color 0.3s;
-}
-
-.error-button:hover {
-  background-color: #a71d2a;
-}
-
-/* Load animation */
-.loader {
-  width: 80px;
-  height: 80px;
-}
-
-.lds-roller,
-.lds-roller div,
-.lds-roller div:after {
-  box-sizing: border-box;
-}
-.lds-roller {
-  display: inline-block;
-  position: relative;
-  width: 80px;
-  height: 80px;
-}
-.lds-roller div {
-  animation: lds-roller 1.2s cubic-bezier(0.5, 0, 0.5, 1) infinite;
-  transform-origin: 40px 40px;
-}
-.lds-roller div:after {
-  content: ' ';
-  display: block;
-  position: absolute;
-  width: 7.2px;
-  height: 7.2px;
-  border-radius: 50%;
-  background: currentColor;
-  margin: -3.6px 0 0 -3.6px;
-}
-.lds-roller div:nth-child(1) {
-  animation-delay: -0.036s;
-}
-.lds-roller div:nth-child(1):after {
-  top: 62.62742px;
-  left: 62.62742px;
-}
-.lds-roller div:nth-child(2) {
-  animation-delay: -0.072s;
-}
-.lds-roller div:nth-child(2):after {
-  top: 67.71281px;
-  left: 56px;
-}
-.lds-roller div:nth-child(3) {
-  animation-delay: -0.108s;
-}
-.lds-roller div:nth-child(3):after {
-  top: 70.90963px;
-  left: 48.28221px;
-}
-.lds-roller div:nth-child(4) {
-  animation-delay: -0.144s;
-}
-.lds-roller div:nth-child(4):after {
-  top: 72px;
-  left: 40px;
-}
-.lds-roller div:nth-child(5) {
-  animation-delay: -0.18s;
-}
-.lds-roller div:nth-child(5):after {
-  top: 70.90963px;
-  left: 31.71779px;
-}
-.lds-roller div:nth-child(6) {
-  animation-delay: -0.216s;
-}
-.lds-roller div:nth-child(6):after {
-  top: 67.71281px;
-  left: 24px;
-}
-.lds-roller div:nth-child(7) {
-  animation-delay: -0.252s;
-}
-.lds-roller div:nth-child(7):after {
-  top: 62.62742px;
-  left: 17.37258px;
-}
-.lds-roller div:nth-child(8) {
-  animation-delay: -0.288s;
-}
-.lds-roller div:nth-child(8):after {
-  top: 56px;
-  left: 12.28719px;
-}
-@keyframes lds-roller {
-  0% {
-    transform: rotate(0deg);
-  }
-  100% {
-    transform: rotate(360deg);
-  }
-}
