+++ src/views/Homepage/Homepage.scss
@@ -1,0 +1,5 @@
+.results {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
